Metadata url and base url as items in the auth context (#33)
Was incorrectly using `path.Join` in order to join URL paths.  This would've failed on windows, and also the redirect for the metadata and user info endpoints ended up being like `http:/blah.okta.com/...` instead of having the correct `//`.

Added a couple functions to the context interface to get these.
Added more todo's around config object to implement the actual use of the discovery endpoints. This will get rid of a lot of the URLs in the config.