BugFix: Timeout should cause the existing execution to abort and these should be finite (#71)
Bug 1; Execution Timeout currently has a bug such that the existing execution is not aborted. Thus the execution continues to run in the background potentially consuming resources.

Bug 2; On a task timeout the UI shows the existing task as running and a new execution begins

Bug 3; Timeouts could cause infinite retries, as it bypassed the retry logic

Solution
Every timeout is treated like a retryable failure, and every retryable failure enforces an abort to be invoked. Abort was already expected to be idempotent, thus invoking it for a terminated execution should not cause side-effects

+improvements
Finalize Child workflow always, even if finalizing parent fails

Unit tests:
Updated unit tests and add new unit tests