Abort AWS Batch Job on Finalize (#58)
AWS Batch Plugin can determine that a task has failed if it detects it can no longer satisfy the failure ratio requirement. When this happens the array task node will be terminated then retried (if there are retry count available). This should trigger a cancellation to the AWS Job as well. 

It was originally written this way to match the old version of this code that, upon discussion, has been deemed a bad design. 

- [X] Need to ensure the cancellation API is idempotent (or add error check/ignore if it's not)
         Validated by running a Workflow and waiting for it to kick off an AWS Job then aborting it. That ended up calling Abort() then Finalize() and the second call was no-op