Add instructions to make minikube's docker env available during 'make docker_build'
Minikube uses a separate docker daemon from the normal system docker
daemon, so you need to tell your shell how to access the minikube docker
daemon. Otherwise when the Task runs it fails immediately because it
tries to pull from Dockerhub.

This also removes a redundant "to" in the following sentence.