Update auth_setup.rst
Update draft based on discussion on Flyte Slack where changing to "http://localhost..." was suggested as a fix. It works for us. Unsure if the comment on using https in case of secure: true is correct. 
Added unfinished explanation for obtaining to ClientSecret, since this is what we're looking for. We tried both internal auth server and external (Google). In the latter case obtaining the ClientSecret was obvious however Google does not support the `grant_type=client_credentials`. It would be good to know if the ClientSecret flow is at all possible with the internal auth server and which external providers acutally support it.