fix(helm): skip redoc when redoc.enabled is set to false
The flyte chart includes four templates: minio, postgres, redis, and redoc.
A property following the pattern `<template_name>.enabled` does exist in
the `values.yaml` file. It is used to enabled/disable the template.

Setting `redoc.enabled` to `false` does not disable the `redoc` deployment
and service. This changeset fixes that problem by following the same
approach with the other templates: wrapping the resoruces under the
`templates/redoc` with an `if` statetement evaluating `Values.redoc.enabled`.

closes 4006

Signed-off-by: cmontemuino <1761056+cmontemuino@users.noreply.github.com>