Support caching subworkflows and launchplans (v2) (#97)
## Overview
Second attempt of https://github.com/unionai/flyte/pull/50. This PR includes stripping launchplan metadata from the `CompiledWorkflowClosure` when computing the workflow digest so that existing workflows with launchplans is not fail registration with this update.

## Test Plan
Same as previous, but added a workflow:
```
@workflow
def parent_wf_compile_test(a: int, b: int):
    my_wf(a=a, b=b)
    my_wf_lp(a=a, b=b)
```
that we registered with a version of propeller without this change and then successfully ran with fast registration (ie. forcing registration attempt) with this change.

## Rollout Plan (if applicable)
See previous PR.

## Upstream Changes
See previous PR.

## Jira Issue
https://unionai.atlassian.net/browse/EXO-46

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [x] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation