Time Partition Granularity and Model Cards (#105)
## Overview
- `artifacts.proto`
  - add model card
  - add granularity to search request
- `artifact_id.proto`
  - define granularity and add to time partition
- Cherry picked commits from https://github.com/flyteorg/flyte/pull/4959

## Test Plan
- Deploy to staging and validate artifacts workflows from the artifact demo repo
- Also validate workflows in our p0 functional tests to insure no interference there

## Rollout Plan (if applicable)
Deploy out to staging tenants and run through above tests, merge in and roll to canary then validate tests there, out to prod

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation