Histogram Bucket Options (#154)
## Overview
Add abstraction to be able to pass buckets custom defined to histogram vectors.

Cherry picked a few changes from OSS into here as well
- [Update spark-on-k8s-operator address in helm charts ](https://github.com/flyteorg/flyte/commit/0b1253bf40e79393078861255cc858f06b22a19e)
- [Fix wrong syntax for path filtering in validate-helm-charts.yaml](https://github.com/flyteorg/flyte/commit/d5784e58226b756719d1f079affe021c85d4646c)

## Test Plan
Unit tests, will be utilizing this in artifacts.

## Rollout Plan (if applicable)
Is currently unused, merging to master and consuming in [this PR](https://github.com/unionai/cloud/pull/6851)

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed

## Jira Issue
https://unionai.atlassian.net/browse/DX-647

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation