Rewrite token endpoint (#167)
## Overview
Add a config to optionally proxy the token URL. Return the updated token URL in the `GetOAuth2Metadata` response.

## Test Plan
- [x] Added a unittest
- [x] Tested e2e as part of https://github.com/unionai/cloud/pull/6881

## Rollout Plan (if applicable)
Disabled by default, so will generally be a noop. We can pick this up in cloud and combine with https://github.com/unionai/cloud/pull/6881 to proxy auth token requests.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed

## Jira Issue
https://unionai.atlassian.net/browse/CLOUD-1690