[Admin/artifacts] Fix parsing of query inputs (#196)
When admin runs an artifact query that has partition values that are parameterized by other inputs, it must first convert those inputs to strings as all partition values are strings.  The parsing for integers was incorrect.

See this for example user code.
https://github.com/unionai/unionai/blob/7e554d70753a3cb0abbb424a143a40a3a26d7f16/tests/functional/workflows/ingest_data.py#L71

The query was coming to the backend without the "depth" partition.