Add vscode settings to repo (#181)
## Overview
This change adds vscode settings to the repo. This may be controversial in OSS, but for the private fork it seems good to use some consistent local settings for linting, etc., in order to spread best practices on the team.

## Test Plan
N/A

## Rollout Plan (if applicable)
N/A

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed