Count accelerated inputs (#227)
## Overview
Add a simple counter to the accelerated input replacement logic so we can gauge how frequently accelerated inputs are used

## Test Plan
- [x] Unit tests pass

## Rollout Plan (if applicable)
Pick up in cloud and deploy to data planes. Start scraping this metric.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed

## Jira Issue
https://unionai.atlassian.net/browse/CLOUD-1777