Add fasttask to PR checks (#271)
## Overview
Plumbs boilerplate to `fasttask/plugin` and wires up to existing checks for lint/unit tests.

## Test Plan
- [x] Run checks workflow from this branch and verify things are integrated and clean

[Lints](https://github.com/unionai/flyte/actions/runs/9083169839/job/24961281752?pr=271) (failing, won't make required for now)
[Generate](https://github.com/unionai/flyte/actions/runs/9083169839/job/24961279800?pr=271)
[Unit Tests](https://github.com/unionai/flyte/actions/runs/9083169839/job/24961284883?pr=271)

## Rollout Plan (if applicable)
N/A

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed

## Issue
fixes EXO-106