fasttask plugin unit tests (#288)
## Overview
This PR adds unit testing for 70.6% of the fasttask plugin go code. Unit tests have been specifically omitted for code that is slated to change (ex. `createPod` and `deletePod` in the `InMemoryBuilder`). In such scenarios the ticket explicitly outlines that unit tests need to be added.

## Test Plan
The plan is to add tests with this PR.

## Rollout Plan (if applicable)
This doesn't need a rollout plan.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed

## Issue
https://linear.app/unionai/issue/EXO-84/fasttask-plugin-unit-tests

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation