Better phase reporting (#296)
## Overview
* Add log lines and improved user facing messaging when worker times out or doesn't heartbeat
* Update phases to be
  * WaitingForResources: no worker available yet
  * Queued: submitted to worker, but not necessarily running yet

## Test Plan
- [x] Updated unittests
- [x] Ran locally and verified states change as expected 

## Rollout Plan (if applicable)
Will bring these to cloud as part of a batch

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed

# Issues
fixes EXO-125
fixes EXO-126