Honor primaryContainerName in default pod spec (#300)
## Overview
Fixes a bug where we weren't propagating primary container name when using the default pod spec.

## Test Plan
Updated unittests to cover this case

## Rollout Plan (if applicable)
Just merge for now. We can take out eventually. No one is depending on this.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed