Added metrics to fasttask builder and plugin (#303)
## Overview
This PR adds prometheus metrics to the fasttask builder and plugin to improve observability of the execution path.

## Test Plan
These metrics were exported and tested locally using a simple `curl` command to the single binary endpoint.

## Rollout Plan (if applicable)
Will need to update the cloud repo to ensure we are correctly pulling these updates.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/EXO-96/fill-out-fasttask-metrics

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation