Override ArrayNode log links with map plugin (#310)
## Overview
This PR adds a configuration option to override ArrayNode log links with those defined in the map plugin. The map plugin contains it's own configuration for log links, which may differ from those defined on the PodPlugin. ArrayNode, executing subNodes as regular tasks (ie. using the PodPlugin) means that it uses the default PodPlugin log templates. The problem, is that live log links in Union are designed to require unique log links for maptasks (which are provided by the map plugin configuration). Therefore, to enable live log links for ArrayNodes, we now support using the mapplugin log links.

It is important to note that all log link template variables are not implemented for ArrayNode as part of this PR, only a superset of those currently used in Union logging.

## Test Plan
This was tested locally and on dogfood, validating that live log links are correctly available.

Validated working: https://dogfood.cloud-staging.union.ai/console/projects/flytesnacks/domains/development/executions/f7c6f87137b3f4b93b41/nodeId/n0/nodes

## Rollout Plan (if applicable)
This may be rolled out immediately.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/EXO-111/support-maptask-specific-log-links-in-arraynode

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation