Building task execution context if not exists on fasttask CheckStatus (#325)
## Overview
This PR fixes an issue where FlytePropeller restarts and the fasttask plugin checks the status of a task execution where the task execution context does not exist. Rather than fail, it creates the task execution context (that will be cleaned up later) and returns running.

## Test Plan
This was tested locally under a variety of failure scenarios.

## Rollout Plan (if applicable)
This may be rolled out immediately.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-1128/fasttask-plugin-checks-status-of-task-execution-failure-on-restart

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation