support adding orphaned pods to existing environments (#328)
## Overview
This PR adds support for managing orphaned pods. This is an extreme corner-case, that may be impossible, but will ensure we do not orphan pods when GCing fasttask environments.

## Test Plan
This was tested locally and added unit tests.

## Rollout Plan (if applicable)
May be rolled out immediately.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-993/support-adding-pods-to-an-orphaned-environment

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [x] Updated documentation