Mount embedded secrets as files (#293)
## Overview
Adds support for mounting secrets as files.

This change is part of series to enable mounting secrets as files:
- flyte fork: https://github.com/unionai/flyte/pull/293 (this change)
- cloud: https://github.com/unionai/cloud/pull/7975
- unionai: https://github.com/unionai/unionai/pull/237

## Test Plan
1. Deploy to dogfood.
2. Create secret using `-f` flag and a file.
3. Run a workflow using the file. Verify that secret value was read correctly.

## Rollout Plan
Run `managed-cluster-<env>-sync-all`.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed

## Issue
ref COR-811

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation