Mount multiple file secrets using single init container (#338)
## Overview
Follow up to #293, in which each secret mounted as file uses a separate init container. This change utilizes a single init container.

## Test Plan
1. Deploy to dogfood.
2. Create multiple secrets using unionai's `-f` flag and file.
3. Run a workflow using the secrets created in previous step.
4. Inspect pod spec.
5. Verify that secret values were read correctly.

## Rollout Plan (if applicable)
Run `managed-cluster-<env>-sync-all`.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
ref COR-811

## Checklist
* [x] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation