[Actor] Fix sandbox makefile and add port (#286)
## Overview
This enables/fixes the local devx that is used for OSS flyte.

Changes
* Update the sandbox-bundled makefile to build a local image and add a port for fast-task to use.
* Update the local config file to turn off artifacts and enable fast-task.
* Add some notes to the readme.

The local endpoint for fast task should be reachable at `localhost:30080` since it should go through the envoy config, but this is as yet untested.


## Test Plan
Local testing only.

## Rollout Plan (if applicable)
Not needed

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If so, please check this box for auditing. Note, this is the responsibility of each developer. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed

## Issue
*TODO: Link Linear issue(s) using [magic words](https://linear.app/docs/github#magic-words). `fixes` will move to merged status, while `ref` will only link the PR.*

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation