add new project state: SYSTEM_ARCHIVED (#351)
## Overview
In https://linear.app/unionai/issue/COR-1437/fix-create-project-calls-for-serverless, we are going to allow Serverless users to register new projects. When we offboard users, we'll archive all their projects, so we need a way to differentiate between archived by users and archived by us.

## Rollout Plan (if applicable)
staging -> canary -> prod

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [x] To be upstreamed to OSS

## Issue
Fixes COR-1437