Gracefully failing to populate log links on propeller restart over fasttask. (#359)
## Overview
Since the fasttask builder stores worker status' in-memory these may be lost on FlytePropeller restart. This causes failures when attempting to lookup a status to populate log links. In this PR, we gracefully fail to populate log links, instead of failing we emit a warn log to facilitate debugging.

## Test Plan
This PR was tested locally under a variety of scenarios.

## Rollout Plan (if applicable)
This may be rolled out immediately.

## Upstream Changes
Should this change be upstreamed to OSS (flyteorg/flyte)? If not, please uncheck this box, which is used for auditing. Note, it is the responsibility of each developer to actually upstream their changes. See [this guide](https://unionai.atlassian.net/wiki/spaces/ENG/pages/447610883/Flyte+-+Union+Cloud+Development+Runbook/#When-are-versions-updated%3F).
- [ ] To be upstreamed to OSS

## Issue
https://linear.app/unionai/issue/COR-1491/fasttask-worker-not-found-in-status-map

## Checklist
* [ ] Added tests
* [ ] Ran a deploy dry run and shared the terraform plan
* [ ] Added logging and metrics
* [ ] Updated [dashboards](https://unionai.grafana.net/dashboards) and [alerts](https://unionai.grafana.net/alerting/list)
* [ ] Updated documentation